+++ Class.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+
+    public class Subject
+    {
+        public string name;
+        public int Score;
+        public int Coefficient;
+    }
