+++ static.cs
@@ -1,0 +1,13 @@
+﻿using System;
+
+
+namespace Record_book
+
+{
+    public class Subject
+    { 
+   public string name;
+        public int Score;
+        public int Coefficient;
+    }
+
